fix data types in CurrencyLayerApiQuotes

+++ CurrencyLayerApiConsoleDemo/CurrencyLayerApiQuotes.cs
@@ -18,7 +18,7 @@
         public float USDBGN { get; set; }
         public float USDBHD { get; set; }
         public float USDBIF { get; set; }
-        public int USDBMD { get; set; }
+        public float USDBMD { get; set; }
         public float USDBND { get; set; }
         public float USDBOB { get; set; }
         public float USDBRL { get; set; }
@@ -27,7 +27,7 @@
         public float USDBTN { get; set; }
         public float USDBWP { get; set; }
         public float USDBYN { get; set; }
-        public int USDBYR { get; set; }
+        public float USDBYR { get; set; }
         public float USDBZD { get; set; }
         public float USDCAD { get; set; }
         public float USDCDF { get; set; }
@@ -35,9 +35,9 @@
         public float USDCLF { get; set; }
         public float USDCLP { get; set; }
         public float USDCNY { get; set; }
-        public int USDCOP { get; set; }
+        public float USDCOP { get; set; }
         public float USDCRC { get; set; }
-        public int USDCUC { get; set; }
+        public float USDCUC { get; set; }
         public float USDCUP { get; set; }
         public float USDCVE { get; set; }
         public float USDCZK { get; set; }
@@ -151,11 +151,11 @@
         public float USDTZS { get; set; }
         public float USDUAH { get; set; }
         public float USDUGX { get; set; }
-        public int USDUSD { get; set; }
+        public float USDUSD { get; set; }
         public float USDUYU { get; set; }
         public float USDUZS { get; set; }
         public float USDVEF { get; set; }
-        public int USDVND { get; set; }
+        public float USDVND { get; set; }
         public float USDVUV { get; set; }
         public float USDWST { get; set; }
         public float USDXAF { get; set; }
